requirements.txt
django-star-ratings